feat(pages): create home and mv index to news
